requirements.txt
# GLASS Data Standardizer v2.0.0 - Production Dependencies
# Pinned versions for reproducibility (tested on 2025-09-28)
# To update: run `pip install --upgrade <package>` and update version here

# Core application dependencies
streamlit==1.39.0
pandas==2.2.3
numpy==2.2.6

# File processing
openpyxl==3.1.5
xlsxwriter==3.2.9
xlrd==2.0.2  # Note: xlrd 2.0+ only supports .xls files, not .xlsx

# Data visualization and analysis
plotly==6.3.0

# AMR Analytics dependencies
kaleido==1.1.0
seaborn==0.13.2
matplotlib==3.10.6

# Text processing and similarity
python-Levenshtein==0.27.1

# System monitoring and performance
psutil==7.0.0
scipy==1.16.2

# API (production)
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
rq==1.16.2
Flask==3.0.2
dash==2.17.1
dash-bootstrap-components==1.6.0